backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-chroma
langchain-community
chromadb
pypdf
tqdm
python-dotenv
sentence-transformers
transformers
torch
